requirements.txt
# =============================================================
# AURA — Automated Unified Review Assistant
# Python dependencies
# =============================================================

# --- Streamlit UI ---
streamlit>=1.30.0

# --- Image Analysis Module (Qdrant multimodal retrieval) ---
openai>=1.10.0
python-dotenv>=1.0.0
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-vector-stores-qdrant>=0.1.0
llama-index-embeddings-openai>=0.1.0
qdrant-client>=1.7.0
Pillow>=9.4.0

# --- Chatbot Module (FAISS RAG) ---
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
requests>=2.31.0

# --- Segmentation Agent (Sa2VA-4B) ---
torch>=2.0.0
numpy>=1.24.0
transformers>=4.40.0
accelerate>=0.25.0
